refactor(game): extract keyboard shortcuts to a hook
First slice of #116 (GameContainer refactor — gravity well at 728 LOC,
22 commits in 14 days at the start of the cycle). The plan there is
to peel self-contained chunks into hooks one PR at a time so each
slice is reviewable and bisectable.

What moved
==========

src/hooks/useKeyboardShortcuts.ts — new file, 131 lines including
comments. Two useEffects:

  1. Undo/redo (Ctrl+Z, Ctrl+Shift+Z, Ctrl+Y) — fires regardless of
     selected cell, gated on PLAYING + !tourOpen.
  2. Cell input (1-9, Backspace/Delete/0, arrows, 'n') — gated on a
     selected cell + PLAYING + !tourOpen.

Both listeners install on window and clean up on unmount. The hook
returns void; it's purely side-effectful.

Why these two together
======================

They're peer-level "global keydown" handlers. Splitting them into two
hooks would just re-create the original boilerplate. Keeping them
co-located in one hook also makes future maintenance simpler — adding
a new shortcut goes in one file.

Why pendingDigitSoundRef + pendingHapticRef are passed in (not owned
=====================================================================
by the hook)

The completion effect in GameContainer also reads these refs to route
sound/haptic on placement. If the hook owned them, the completion
effect would have to receive them back somehow, which is more
plumbing than passing them in. Refs are stable identity, so passing
them through doesn't perturb dep arrays.

What didn't change
==================

- Behavior — every keydown path is byte-identical with the original
  inline implementation. Same gating, same dep arrays for the effects.
- The completion effect — still reads pendingDigitSoundRef /
  pendingHapticRef from GameContainer.
- Tests — Board.test.tsx exercises some of this via fireEvent.keyDown
  on cells, all 10 tests still pass.

Verified
========

- typecheck: 0 errors
- lint: clean
- Full unit suite passes
- GameContainer.tsx: 728 → 652 LOC (-76)
- New file: 131 LOC (mostly comments + the two effects' bodies)

Next slices (separate PRs)
==========================

- usePuzzleLifecycle: handleStartDaily / handleNewGame /
  handleDifficulty / handleType + completion-effect daily branch
- useShare: handleShare + ShareToast state
- useAutoStartIdle: the IDLE → newGame effect with deep-link parsing
